moving tests and creating set up guides
