+++ frontend/vitest.config.ts
@@ -6,7 +6,8 @@ export default defineConfig({
   plugins: [vue()],
   test: {
     environment: 'jsdom',
-    globals: true
+    globals: true,
+    include: ['tests/**/*.test.ts']
   },
   resolve: {
     alias: {
